requirements.txt
click==6.6
cuid.py==0.2
Flask==0.11.1
Flask-Cors==2.1.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
psycopg2==2.6.2
python-dateutil==2.5.3
requests==2.10.0
six==1.10.0
Werkzeug==0.11.10
